requirements.txt
flask>=3.0
flask-cors>=4.0
transformers>=4.40
torch>=2.0
Pillow>=10.0
